release 1.13

+++ z80sim/simfun.c
@@ -1,7 +1,7 @@
 /*
  * Z80SIM  -  a	Z80-CPU	simulator
  *
- * Copyright (C) 1987-2006 by Udo Munk
+ * Copyright (C) 1987-2007 by Udo Munk
  *
  * History:
  * 28-SEP-87 Development on TARGON/35 with AT&T Unix System V.3
@@ -13,11 +13,12 @@
  * 10-JUN-92 Release 1.6  long casting problem solved with COHERENT 3.2
  *			  and some optimization
  * 25-JUN-92 Release 1.7  comments in english and ported to COHERENT 4.0
- * 04-OCT-06 Release 1.8  modified to compile on modern POSIX OS's
+ * 02-OCT-06 Release 1.8  modified to compile on modern POSIX OS's
  * 18-NOV-06 Release 1.9  modified to work with CP/M sources
  * 08-DEC-06 Release 1.10 modified MMU for working with CP/NET
  * 17-DEC-06 Release 1.11 TCP/IP sockets for CP/NET
- * 25-DEC-06 Release 1.12 CPU speed option and 100 ticks interrupt
+ * 25-DEC-06 Release 1.12 CPU speed option
+ * 19-FEB-07 Release 1.13 various improvements
  */
 
 /*
@@ -37,12 +38,12 @@ int exatoi(char *str)
 {
 	register int num = 0;
 
-	while (isxdigit(*str)) {
+	while (isxdigit((int)*str)) {
 		num *= 16;
 		if (*str <= '9')
 			num += *str - '0';
 		else
-			num += toupper(*str) - '7';
+			num += toupper((int)*str) - '7';
 		str++;
 	}
 	return(num);
